requirements.txt
cudf-cu12==24.8.2
cuml-cu12==24.8.0
dgl==2.2.1+cu121
einops==0.8.0
lightning==2.3.3
ogb==1.3.6
rmm-cu12==24.8.2
scikit-learn==1.5.0
scipy==1.13.1
tensorboardX==2.6.2.2
torch==2.3.0
torch_cluster==1.6.3+pt23cu121
torch_geometric==2.5.3
torch_scatter==2.1.2+pt23cu121
torch_sparse==0.6.18+pt23cu121
tqdm==4.66.4
